refactor: refactor ddl procedure
