+++ src/common/meta/src/ddl/drop_table.rs
@@ -14,17 +14,8 @@
 
 use api::v1::region::{region_request, DropRequest as PbDropRegionRequest};
 use async_trait::async_trait;
-use client::region::RegionRequester;
 use common_error::ext::ErrorExt;
 use common_error::status_code::StatusCode;
-use common_meta::cache_invalidator::Context;
-use common_meta::ident::TableIdent;
-use common_meta::key::table_info::TableInfoValue;
-use common_meta::key::table_name::TableNameKey;
-use common_meta::key::table_route::TableRouteValue;
-use common_meta::rpc::ddl::DropTableTask;
-use common_meta::rpc::router::{find_leader_regions, find_leaders, RegionRoute};
-use common_meta::table_name::TableName;
 use common_procedure::error::{FromJsonSnafu, ToJsonSnafu};
 use common_procedure::{
     Context as ProcedureContext, LockKey, Procedure, Result as ProcedureResult, Status,
@@ -39,20 +30,29 @@ use table::engine::TableReference;
 use table::metadata::{RawTableInfo, TableId};
 
 use super::utils::handle_retry_error;
+use crate::cache_invalidator::Context;
+use crate::ddl::utils::handle_operate_region_error;
 use crate::ddl::DdlContext;
-use crate::error::{self, Result, TableMetadataManagerSnafu};
+use crate::error::{self, Result};
+use crate::ident::TableIdent;
+use crate::key::table_info::TableInfoValue;
+use crate::key::table_name::TableNameKey;
+use crate::key::table_route::TableRouteValue;
 use crate::metrics;
-use crate::procedure::utils::handle_request_datanode_error;
+use crate::rpc::ddl::DropTableTask;
+use crate::rpc::router::{find_leader_regions, find_leaders, RegionRoute};
+use crate::table_name::TableName;
 
 pub struct DropTableProcedure {
-    context: DdlContext,
-    data: DropTableData,
+    pub context: DdlContext,
+    pub data: DropTableData,
 }
 
+#[allow(dead_code)]
 impl DropTableProcedure {
-    pub(crate) const TYPE_NAME: &'static str = "metasrv-procedure::DropTable";
+    pub const TYPE_NAME: &'static str = "metasrv-procedure::DropTable";
 
-    pub(crate) fn new(
+    pub fn new(
         cluster_id: u64,
         task: DropTableTask,
         table_route_value: TableRouteValue,
@@ -65,7 +65,7 @@ impl DropTableProcedure {
         }
     }
 
-    pub(crate) fn from_json(json: &str, context: DdlContext) -> ProcedureResult<Self> {
+    pub fn from_json(json: &str, context: DdlContext) -> ProcedureResult<Self> {
         let data = serde_json::from_str(json).context(FromJsonSnafu)?;
         Ok(Self { context, data })
     }
@@ -82,13 +82,12 @@ impl DropTableProcedure {
                 table_ref.schema,
                 table_ref.table,
             ))
-            .await
-            .context(TableMetadataManagerSnafu)?;
+            .await?;
 
         ensure!(
             exist,
             error::TableNotFoundSnafu {
-                name: table_ref.to_string()
+                table_name: table_ref.to_string()
             }
         );
 
@@ -106,8 +105,7 @@ impl DropTableProcedure {
 
         table_metadata_manager
             .delete_table_metadata(table_info_value, table_route_value)
-            .await
-            .context(error::TableMetadataManagerSnafu)?;
+            .await?;
 
         info!("Deleted table metadata for table {table_id}");
 
@@ -137,15 +135,14 @@ impl DropTableProcedure {
                 },
                 table_ident,
             )
-            .await
-            .context(error::InvalidateTableCacheSnafu)?;
+            .await?;
 
         self.data.state = DropTableState::DatanodeDropRegions;
 
         Ok(Status::executing(true))
     }
 
-    async fn on_datanode_drop_regions(&self) -> Result<Status> {
+    pub async fn on_datanode_drop_regions(&self) -> Result<Status> {
         let table_id = self.data.table_id();
 
         let region_routes = &self.data.region_routes();
@@ -153,7 +150,7 @@ impl DropTableProcedure {
         let mut drop_region_tasks = Vec::with_capacity(leaders.len());
 
         for datanode in leaders {
-            let clients = self.context.datanode_clients.clone();
+            let clients = self.context.datanode_manager.clone();
 
             let regions = find_leader_regions(region_routes, &datanode);
             let region_ids = regions
@@ -169,12 +166,11 @@ impl DropTableProcedure {
                         region_id: region_id.as_u64(),
                     });
 
-                    let client = clients.get_client(&datanode).await;
-                    let requester = RegionRequester::new(client);
+                    let requester = clients.datanode(&datanode).await;
 
                     if let Err(err) = requester.handle(request).await {
                         if err.status_code() != StatusCode::RegionNotFound {
-                            return Err(handle_request_datanode_error(datanode)(err));
+                            return Err(handle_operate_region_error(datanode)(err));
                         }
                     }
                 }
@@ -232,11 +228,11 @@ impl Procedure for DropTableProcedure {
 
 #[derive(Debug, Serialize, Deserialize)]
 pub struct DropTableData {
-    state: DropTableState,
-    cluster_id: u64,
-    task: DropTableTask,
-    table_route_value: TableRouteValue,
-    table_info_value: TableInfoValue,
+    pub state: DropTableState,
+    pub cluster_id: u64,
+    pub task: DropTableTask,
+    pub table_route_value: TableRouteValue,
+    pub table_info_value: TableInfoValue,
 }
 
 impl DropTableData {
@@ -277,7 +273,7 @@ impl DropTableData {
 }
 
 #[derive(Debug, Serialize, Deserialize, AsRefStr)]
-enum DropTableState {
+pub enum DropTableState {
     /// Prepares to drop the table
     Prepare,
     /// Removes metadata
@@ -286,72 +282,4 @@ enum DropTableState {
     InvalidateTableCache,
     /// Drops regions on Datanode
     DatanodeDropRegions,
-}
-
-#[cfg(test)]
-mod test {
-    use std::collections::HashSet;
-    use std::sync::{Arc, Mutex};
-
-    use super::*;
-    use crate::procedure::utils::mock::EchoRegionServer;
-    use crate::procedure::utils::test_data;
-
-    #[tokio::test]
-    async fn test_on_datanode_drop_regions() {
-        let drop_table_task = DropTableTask {
-            catalog: "my_catalog".to_string(),
-            schema: "my_schema".to_string(),
-            table: "my_table".to_string(),
-            table_id: 42,
-        };
-        let procedure = DropTableProcedure::new(
-            1,
-            drop_table_task,
-            TableRouteValue::new(test_data::new_region_routes()),
-            TableInfoValue::new(test_data::new_table_info()),
-            test_data::new_ddl_context(),
-        );
-
-        let (region_server, mut rx) = EchoRegionServer::new();
-
-        let datanodes = find_leaders(&procedure.data.table_route_value.region_routes);
-        for peer in datanodes {
-            let client = region_server.new_client(&peer);
-            procedure
-                .context
-                .datanode_clients
-                .insert_client(peer, client)
-                .await;
-        }
-
-        let expected_dropped_regions = Arc::new(Mutex::new(HashSet::from([
-            RegionId::new(42, 1),
-            RegionId::new(42, 2),
-            RegionId::new(42, 3),
-        ])));
-        let handle = tokio::spawn({
-            let expected_dropped_regions = expected_dropped_regions.clone();
-            let mut max_recv = expected_dropped_regions.lock().unwrap().len();
-            async move {
-                while let Some(region_request::Body::Drop(request)) = rx.recv().await {
-                    let region_id = RegionId::from_u64(request.region_id);
-
-                    expected_dropped_regions.lock().unwrap().remove(&region_id);
-
-                    max_recv -= 1;
-                    if max_recv == 0 {
-                        break;
-                    }
-                }
-            }
-        });
-
-        let status = procedure.on_datanode_drop_regions().await.unwrap();
-        assert!(matches!(status, Status::Done));
-
-        handle.await.unwrap();
-
-        assert!(expected_dropped_regions.lock().unwrap().is_empty());
-    }
 }
