+++ src/common/meta/src/lib.rs
@@ -16,6 +16,7 @@
 
 pub mod cache_invalidator;
 pub mod datanode_manager;
+pub mod ddl;
 pub mod ddl_manager;
 pub mod error;
 pub mod heartbeat;
