+++ src/meta-srv/src/procedure.rs
@@ -12,9 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-pub mod alter_table;
-pub mod create_table;
-pub mod drop_table;
 pub mod region_failover;
 pub(crate) mod state_store;
+#[cfg(test)]
+mod tests;
 mod utils;
